requirements.txt
#remove the version number from the requirements.txt file
alembic
amqp
annotated-types
anyio
bcrypt==3.1.7
billiard
celery
cffi
python-jose 
click
click-didyoumean
click-plugins
click-repl
colorama
dnspython
email_validator
fastapi
greenlet
h11
idna
itsdangerous
kombu
Mako
MarkupSafe
packaging
passlib
prompt_toolkit
psycopg2-binary
pycparser
pydantic
pydantic_core
pydantic-settings
python-dateutil
redis
six
sniffio
SQLAlchemy
starlette
typing-inspection
typing_extensions
tzdata
uvicorn
vine
wcwidth